Add unique color index for use in rasters
