+++ lib/io.js
@@ -16,17 +16,17 @@ function fetchJson(url) {
   return JSON.parse(data)
 }
 
-function lesKildedatafil(filename) {
+function lesKildedatafil(filename, defaultJson) {
   const jsonPath = config.kildedataPath + "/" + filename + ".json"
-  return readJson(jsonPath)
+  return readJson(jsonPath, defaultJson)
 }
 
-function lesDatafil(filename, extension) {
+function lesDatafil(filename, extension, defaultJson) {
   const jsonPath = config.getDataPath(filename, extension)
-  return readJson(jsonPath)
+  return readJson(jsonPath, defaultJson)
 }
 
-function skriveKildedatafil(filename) {
+function skrivKildedatafil(filename, o) {
   const jsonPath = config.kildedataPath + "/" + filename + ".json"
   return writeJson(jsonPath, o)
 }
@@ -60,8 +60,10 @@ function skrivLoggLinje(aksjon, filePath, json) {
   )
 }
 
-function readJson(filePath) {
+// Read file, parse to JSON.  If file doesn't exists defaultJson will be returned
+function readJson(filePath, defaultJson) {
   log.info("Åpner " + filePath)
+  if (defaultJson && !fs.existsSync(filePath)) return defaultJson
   let data = fs.readFileSync(filePath, "utf8")
   //  data = data.replace(/^\uFEFF/, '') // node #fail https://github.com/nodejs/node/issues/6924
   if (data.charCodeAt(0) === 0xfeff) data = data.slice(1)
@@ -131,6 +133,7 @@ module.exports = {
   readJson,
   writeBinary,
   writeJson,
+  skrivKildedatafil,
   skrivDatafil,
   skrivBuildfil,
   fetchJson,
